#6 [feat] 도메인 엔티티 작성
