+++ module-api/src/main/java/com/mile/PostController.java
@@ -1,0 +1,4 @@
+package com.mile;
+
+public class PostController {
+}
