--- module-api/src/main/java/com/mile/post/PostController.java
@@ -1,4 +1,0 @@
-package com.mile.post;
-
-public class PostController {
-}
