+++ module-domain/src/main/java/com/mile/curious/domain/Curious.java
@@ -1,0 +1,23 @@
+package com.mile.curious.domain;
+
+import com.mile.post.Post;
+import com.mile.config.BaseTimeEntity;
+import com.mile.user.domain.User;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class Curious extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    private Post post;
+
+    @ManyToOne
+    private User user;
+}
