+++ module-domain/src/main/java/com/mile/curious/serivce/CuriousService.java
