+++ module-domain/src/main/java/com/mile/moim/domain/Moim.java
@@ -1,0 +1,21 @@
+package com.mile.moim.domain;
+
+import com.mile.user.domain.User;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class Moim {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private String id;
+    @ManyToOne
+    private User owner;
+    private String name;
+    private String image;
+    private String information;
+    private boolean isPublic;
+}
