--- module-domain/src/main/java/com/mile/post/service/PostService.java
@@ -1,7 +1,0 @@
-package com.mile.post.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PostService {
-}
